% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstat.R
\name{clx_stat}
\alias{clx_stat}
\title{Cai--Liu--Xia max-type covariance statistic (C back-end)}
\usage{
clx_stat(XA, XB)
}
\arguments{
\item{XA, XB}{numeric data matrices with the same number of columns.}
}
\value{
the maximum standardised squared difference of covariance entries.
}
\description{
Cai--Liu--Xia max-type covariance statistic (C back-end)
}
